+++ persistence/migrations/0120_userSessions.ts
@@ -17,7 +17,7 @@ export const up = async (knex: Knex): Promise<void> =>
 
     table.boolean("deleted");
     table.bigInteger("createdAt");
-    table.integer("updatedAt");
+    table.bigInteger("updatedAt");
 
     table.foreign("userId").references("id").inTable("users");
